+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Crawler 9.1.5-dev
 
+### Added
+* Error when Page 0 is used for crawler:buildQueue
+
+
 ### Change
 * Removed all references to cHash as it is not used anymore.